Move main plotting functionality to importable functions
